--- a/Misc/NEWS.d/next/Core and Builtins/2020-03-12-02-41-12.bpo-37207.ye7OM3.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2020-03-12-02-41-12.bpo-37207.ye7OM3.rst
@@ -1,2 +1,2 @@
-Speed up calls to ``tuple()`` by using the PEP 590 ``vectorcall`` calling
+Speed up calls to ``tuple()`` by using the :pep:`590` ``vectorcall`` calling
 convention. Patch by Dong-hee Na.