Master
```
./python.exe -m pyperf timeit "tuple()"
.....................
Mean +- std dev: 262 ns +- 9 ns

./python.exe -m pyperf timeit "tuple((1, 2, 3, 4, 5))"
.....................
Mean +- std dev: 361 ns +- 15 ns
./python.exe -m pyperf timeit "tuple([1, 2, 3, 4, 5])"
.....................
Mean +- std dev: 927 ns +- 29 ns
```
PEP-590
```
./python.exe -m pyperf timeit "tuple()"
.....................
Mean +- std dev: 176 ns +- 11 ns

./python.exe -m pyperf timeit "tuple((1, 2, 3, 4, 5))"
.....................
Mean +- std dev: 203 ns +- 13 ns

./python.exe -m pyperf timeit "tuple([1, 2, 3, 4, 5])"
.....................
Mean +- std dev: 714 ns +- 35 ns
```

<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
bpo-NNNN: Summary of the changes made
```

Where: bpo-NNNN refers to the issue number in the https://bugs.python.org.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `master`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `master`.

-->


<!-- issue-number: [bpo-37207](https://bugs.python.org/issue37207) -->
https://bugs.python.org/issue37207
<!-- /issue-number -->